refactor chord with data-driven factories, extract progression and arpeggio
- CHORD_TYPES dict + Chord.build() replaces duplicated interval logic
- All create_* classmethods preserved as one-liner wrappers
- ChordProgression extracted to composition/progression.py
- Arpeggio + ArpeggioPattern extracted to composition/arpeggio.py
- Zero API breakage, zero test logic changes
