+++ midigen/composition/__init__.py
@@ -1,1 +1,5 @@
 """Composition layer — depends only on theory/."""
+
+from midigen.composition.chord import Chord, CHORD_TYPES
+from midigen.composition.progression import ChordProgression
+from midigen.composition.arpeggio import Arpeggio, ArpeggioPattern
